fix: 手写请求结构数字字段全形态 Flex 统一——新增切片/map 宽松类型（下游反馈：Go 严格绑定 vs PHP 弱类型）
- validator 新增 FlexInt32Slice []int32（接受 [1,2]/["1","2"]/混合/空/null，非法元素报错含索引）与 FlexInt32Map map[int32]int（JSON 键恒为字符串——正是严格 Go 失败点；键/值均宽松，非法报错含键信息）；未加无调用方的 FlexInt64Slice/FlexFloat64Slice/FlexInt32Ptr（红线：无调用方不加）
- 8 处手写请求字段转 Flex（json tag/binding 标签不变）：admin_group.Pid→FlexInt32、Rules []int32→FlexInt32Slice；user.AdminID→FlexInt32、JoinTime→FlexInt64；routine_config.Weigh→FlexInt32；admin.NullableParentID.Value *int32→*FlexInt32（null→nil 语义不变）；crud.SyncIDs map[int32]int→FlexInt32Map；dto/common.go IDS.ID→FlexInt32（del 批量主键）
- 消费链最小显式转换（无 helper、不改签名）：admin_group.go []int32(params.Rules)/int32(params.ID)；user.go int64(params.JoinTime)/int32(params.ID)；routine_config.go int32(params.Weigh)；admin.go *FlexInt32→*int32 指针解引用（局部变量）；crud.go int32(param.ID)/map[int32]int(params.SyncIDs)；IDS 消费方 admin_rule/security_data_recycle/security_sensitive_data/admin 各加 int32(params.ID)；copier.Copy 按既有模式免改
- 测试：flex_number_test 新增 4 组（切片 7 形态/非法元素、map 6 形态/非法键值）；新增 handler/flex_binding_test.go（无类型声明，豁免命名执法）——AdminGroup/User/Config/Admin/crud 结构级字符串形态解码回归 + IDS 经 gin ShouldBindJSON 验证 {"id":"7"} 成功、缺 id 仍 required（required 语义无回归）
- 验证：build + 308 测试（validator 129 含新增 17 用例）+ boundary 执法 + 命名执法全绿

+++ internal/admin/handler/admin_rule.go
@@ -122,7 +122,7 @@ func (h *AdminRuleHandler) Edit(ctx *gin.Context) {
 		response.FailByErr(ctx, validator.GetError(params, err))
 		return
 	}
-	adminRule, err := h.adminRuleM.GetOne(ctx, params.ID)
+	adminRule, err := h.adminRuleM.GetOne(ctx, int32(params.ID))
 	if err != nil {
 		response.FailByErr(ctx, err)
 		return
